Add undo functionality
The undo functionality has been added. This acts
as en erase button, more than an actual undo, in
that it will erase what is on the screen but not
go back to previous operations or expressions
that were on the screen before (i.e. there is
no history, only what is on the screen).